Correção KM veiculo

+++ Api/Model/VeiculoModel.cs
@@ -28,7 +28,7 @@ namespace Api.Model
 
         public int? CapacidadeTanque { get; set; }
 
-        public int? Kilometragem { get; set; }
+        public double? Kilometragem { get; set; }
 
         public int? AnoFabricacao { get; set; }
